Use S3 image bucket if configured for uploaded images
Expose a new method which generates a pre-signed S3 upload to the S3
image bucket if an image bucket is configured. Then on the client side,
when uploading an image to a spreadsheet, attempt to get a pre-signed S3
URL. If that works, upload the image and insert the image by URL into
the spreadsheet. If it doesn't work, fall back to the old method of
uploading the image to Google Sheets and inserting it as a blob.
