+++ package.json
@@ -20,6 +20,7 @@
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.485.0",
+    "@aws-sdk/s3-presigned-post": "^3.485.0",
     "@babel/runtime": "^7.23.7",
     "@bugsnag/js": "^7.22.3",
     "@bugsnag/plugin-react": "^7.19.0",
